Major re-design of MLPlan builders. Also removed annyoing split-warning

+++ JAICore/jaicore-ml/src/main/java/ai/libs/jaicore/ml/regression/loss/dataset/MeanSquaredError.java
@@ -14,7 +14,7 @@ public class MeanSquaredError extends AUnboundedRegressionMeasure {
 	@Override
 	public double loss(final List<? extends Double> expected, final List<? extends Double> actual) {
 		this.checkConsistency(expected, actual);
-		return StatisticsUtil.mean(IntStream.range(0, expected.size()).mapToObj(x -> Double.valueOf(SQUARED_ERROR_LOSS.loss(expected.get(x), actual.get(x)))).collect(Collectors.toList()));
+		return StatisticsUtil.mean(IntStream.range(0, expected.size()).mapToObj(x -> Double.valueOf(SQUARED_ERROR_LOSS.loss((double)expected.get(x), (double)actual.get(x)))).collect(Collectors.toList()));
 	}
 
 }
